+++ spec/support/apps/rails_4.0_app/config/routes.rb
@@ -1,4 +1,5 @@
 Rails40App::Application.routes.draw do
   devise_for :users
+  mount MyAppWrapped => "/rack_test"
   root :to => "home#index"
 end